requirements.txt
# Core
sqlalchemy>=1.4.23
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Database
alembic>=1.7.4

# Frontend
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.1.1
openai>=1.3.0
requests>=2.31.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Development
autopep8>=1.6.0
flake8>=4.0.1
black>=22.3.0
isort>=5.10.1
